Migrating GraphQL Integration Tests
